+ 2D renderer for simpler API in 2D games, currently draws Quads with optional colour and optional texture + uses single combined shader for all draws + Default single white pixel texture
+ Set up sandbox to have fully functional test menu at start up with options to load number of tests and scenes and correctly dispose of them when they switching away

+++ Sandbox/res/shaders/Batch.vert
@@ -3,7 +3,7 @@
 layout(location = 0) in vec4 position;
 layout(location = 1) in vec2 texCoords;
 layout(location = 2) in vec4 colour;
-layout(location = 3) in int texIndex;
+layout(location = 3) in float texIndex;
 
 out vec2 v_TexCoords;
 out vec4 v_Colour;
@@ -16,5 +16,5 @@ void main() {
 	gl_Position = u_VP * u_Transform * position;
 	v_TexCoords = texCoords;
 	v_Colour = colour;
-	v_TexIndex = texIndex;
+	v_TexIndex = int(texIndex);
 }
